Fix broken import of non-exported file
