+++ packages/nimble-components/src/rich-text-mention/users/types.ts
@@ -1,0 +1,7 @@
+import type { RichTextMentionUsers } from '.';
+import type { MappingUser } from '../../mapping/user';
+
+export interface UserMentionElements {
+    userMentionElement: RichTextMentionUsers;
+    mappingElements: MappingUser[];
+}
