+++ packages/nimble-components/src/rich-text/editor/testing/types.ts
@@ -4,9 +4,6 @@ import {
     keyArrowRight,
     keyArrowUp
 } from '@microsoft/fast-web-utilities';
-import type { MappingUser } from '../../../mapping/user';
-import type { RichTextMentionTest } from '../../../rich-text-mention/base/tests/rich-text-mention.fixtures';
-import type { RichTextMentionUsers } from '../../../rich-text-mention/users';
 
 export const ToolbarButton = {
     bold: 0,
@@ -32,16 +29,6 @@ export type LabelProvider =
     | 'toggleNumberedList';
 
 export type ToolbarButtonKey = keyof typeof ToolbarButton;
-
-export interface UserMentionElements {
-    userMentionElement: RichTextMentionUsers;
-    mappingElements: MappingUser[];
-}
-
-export interface TestMentionElements {
-    testMentionElement: RichTextMentionTest;
-    mappingElements: MappingUser[];
-}
 
 export interface MappingConfiguration {
     key?: string;
